create object to manage gdb connection
